% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{logisticregsummary}
\alias{logisticregsummary}
\title{Summarize the results from a logistic regression analysis}
\usage{
logisticregsummary(
  glmobject,
  odds_ratio = T,
  param = T,
  conf_intervals = F,
  vif_factors = F
)
}
\arguments{
\item{glmobject}{a fitted regression model from \code{glm}.}

\item{param}{\code{TRUE} if parameter estimates, standard errors etc is computed.}

\item{conf_intervals}{\code{TRUE} if confidence intervals for parameters.}

\item{vif_factors}{\code{TRUE} if variance inflation factors are to be printed.}
}
\value{
list with two tables: param, odds_ratio
}
\description{
Alternative to \code{summary.glm} to summarize a regression from \code{glm}.
Prints a table similar to the one generated by SAS and Minitab.
}
\examples{
library(regkurs)
glmfit <- glm(survived ~ age + sex + firstclass, data = titanic, family = binomial)
logisticregsummary(glmfit)
}
