% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_analysis.R, R/simple resid anal.R
\name{res.diagnostics}
\alias{res.diagnostics}
\title{Plots residual diagnostics for regression model.}
\usage{
res.diagnostics(lm_obj)

res.diagnostics(lm_obj)
}
\arguments{
\item{lm_obj}{a fitted regression model from \code{lm}.}
}
\value{
list with six plot handles

list with at least six plots
}
\description{
Plots residual diagnostics for regression model.

Plots residual diagnostics for regression model.
}
\examples{
library(regkurs)
lmfit = lm(nRides ~ temp + hum + windspeed, data = bike)
res.diagnostics(lmfit)
library(regkurs)
lmfit = lm(nRides ~ temp + hum + windspeed, data = bike)
res.diagnostics(lmfit)
}
